# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation for the Project",
    "version": "1.0.0",
    "title": "Project API"
  },
...
  "paths": {
...
    "/user/:board_id": {
      "get": {
        "tags": ["User"],
        "summary": "Get all users by board ID",
        "description": "Fetches all users participating in a specific board",
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users"
          },
          "500": {
            "description": "Users in board search failed"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": ["User"],
        "summary": "Get all users",
        "description": "Fetches all users in the system",
        "responses": {
          "200": {
            "description": "A list of users"
          },
          "500": {
            "description": "Users search failed"
          }
        }
      }
    },
    "/user/:id": {
      "post": {
        "tags": ["User"],
        "summary": "Update a user",
        "description": "Updates the information of a specific user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "User updated successfully"
          },
          "500": {
            "description": "User update failed"
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete a user",
        "description": "Deletes the information of a specific user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "User deleted successfully"
          },
          "500": {
            "description": "User update failed"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format **Bearer &lt;token&gt;**"
    }
  }
}